fix(vault): Correct potential problems on 32-bit systems (refs #6181)
The vault's id generation algorithm has been changed to fit the capacity
of the system:
- On 64-bit systems, vault's id will be 64-bit positive integers
  (int8). During migration, old ids will be regenerated with values
  above between 2^32 and PHP_INT_MAX.
- On 32-bit systems, vault's id will be 32-bit positive integers
  (int4). During migration, old ids will be regenerated between the
  current seq_id_vaultdiskstorage's value and PHP_INT_MAX.

From now on, vault's id should be treated as string to prevent integer
overflows problems.

Constructions that can leads to unexpected results are (but not limited
to):
- Casts with intval() function
- Formatting with printf()-style functions (e.g., sprintf("%d", $value))
- Arithmetic comparison and operation

+++ App/Dav/dav-pg.sql
@@ -46,7 +46,7 @@ create index properties_path on  dav.properties(path);
 CREATE TABLE dav.sessions (
   session varchar(200) NOT NULL ,
   owner varchar(200) NOT NULL ,
-  vid int NOT NULL ,
+  vid bigint NOT NULL ,
   fid int NOT NULL ,
   expires int NOT NULL default '0',
   PRIMARY KEY  (session)
